+++ jwst/datamodels/contrast.py
@@ -1,3 +1,0 @@
-from stdatamodels.jwst.datamodels.contrast import ContrastModel
-
-__all__ = ['ContrastModel']