+++ jwst/datamodels/guider.py
@@ -1,3 +1,0 @@
-from stdatamodels.jwst.datamodels.guider import GuiderRawModel, GuiderCalModel
-
-__all__ = ['GuiderRawModel', 'GuiderCalModel']