+++ jwst/datamodels/multiextract1d.py
@@ -1,3 +1,0 @@
-from stdatamodels.jwst.datamodels.multiextract1d import MultiExtract1dImageModel
-
-__all__ = ['MultiExtract1dImageModel']