+++ jwst/datamodels/nirspec_flat.py
@@ -1,3 +1,0 @@
-from stdatamodels.jwst.datamodels.nirspec_flat import NirspecFlatModel, NirspecQuadFlatModel
-
-__all__ = ['NirspecFlatModel', 'NirspecQuadFlatModel']