+++ jwst/datamodels/rscd.py
@@ -1,3 +1,0 @@
-from stdatamodels.jwst.datamodels.rscd import RSCDModel
-
-__all__ = ['RSCDModel']