+++ jwst/datamodels/spec.py
@@ -1,3 +1,0 @@
-from stdatamodels.jwst.datamodels.spec import SpecModel
-
-__all__ = ['SpecModel']